Apex Small fix / Group_Common compute

+++ basis/bin/destroy_all.sh
@@ -34,6 +34,8 @@ elif [ "$TF_VAR_deploy_strategy" == "function" ]; then
   oci os object bulk-delete -bn ${TF_VAR_prefix}-public-bucket --force
 fi
 
-title "Terraform Destroy"
-src/terraform/destroy.sh --auto-approve -no-color
+if [ -f $TARGET_DIR/terraform.tfstate ]; then
+  title "Terraform Destroy"
+  src/terraform/destroy.sh --auto-approve -no-color
+fi
 echo "Destroy time: ${SECONDS} secs"
